Moved iced logic out of main into user_interface
