+++ src/client/gui/bb_tab/login.rs
@@ -2,8 +2,7 @@ use iced::Element;
 use iced::widget::{container, text, text_input, Column, Space};
 use iced_core::Length::Fill;
 use iced_core::Theme;
-use crate::{Message, UserInterface};
-use crate::client::backend::login_state::LoginState;
+use crate::client::gui::user_interface::{Message, UserInterface};
 use crate::client::gui::app::App;
 use crate::client::gui::bb_theme::color;
 use crate::client::gui::bb_theme::color::{ERROR_COLOR, TEXT_COLOR};
