+++ src/client/gui/bb_widget/activity/activity.rs
@@ -18,7 +18,7 @@ use crate::client::gui::bb_theme::custom_button::{ButtonStyle, DEFAULT_BUTTON_RA
 use crate::client::gui::bb_widget::activity::date_utils::{get_date_by_offset, get_end_dates_of_offsets, get_start_dates_of_offsets, started_weeks_in_period, DateScope, Offset, DAYS_PER_WEEK};
 use crate::client::gui::bb_widget::widget_utils::INDENT;
 use crate::client::gui::mascots::Mascot;
-use crate::Message;
+use crate::client::gui::user_interface::Message;
 
 const DEFAULT_NAVIGATION_BUTTON_WIDTH: f32 = 130.0;
 const DEFAULT_NAVIGATION_BUTTON_HEIGHT: f32 = 40.0;
